Add --simple-io option for subprocesses and break out console.h and cpp
